feat: add useGenreAll hook for fetching all genres and update CreateProduct form
- Implemented `useGenreAll` hook to fetch all genres from the API.
- Refactored `CreateProduct` component to utilize `useGenreAll` instead of `useGenre`.
- Updated form fields in `CreateProduct` to match new product structure.
- Added new fields for original price and quantity in the product form.
- Enhanced image upload functionality with file and URL options.
- Created `ProductResponse`, `ProductMapper`, `ProductRepository`, and `ProductService` for backend product management.
- Implemented `ProductServiceImpl` with basic findAll functionality.

+++ backend/src/main/java/com/dev/backend/entity/Product.java
@@ -26,11 +26,11 @@ public class Product {
     private Integer id;
     private String name;
     private Integer quantity;
-    private BigDecimal originalPrice;
-    private BigDecimal price;
+    private Integer originalPrice;
+    private Integer price;
 
     private String description;
-    private int weight;
+    private Integer weight;
 
     private String isbn;
